PIC-2929 conditionally supply random operationID when running locally
