+++ src/main/kotlin/uk/gov/justice/digital/hmpps/crimeportalgateway/application/LocalTelemetryConfig.kt
@@ -1,0 +1,14 @@
+package uk.gov.justice.digital.hmpps.crimeportalgateway.application
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Profile
+import java.util.UUID
+
+@Profile(value = ["test", "local"])
+class LocalTelemetryConfig : BaseTelemetryConfig() {
+
+    @Bean
+    fun getOperationId(): () -> String? {
+        return { UUID.randomUUID().toString() }
+    }
+}
