+++ src/main/kotlin/uk/gov/justice/digital/hmpps/crimeportalgateway/application/TelemetryConfig.kt
@@ -1,39 +1,17 @@
 package uk.gov.justice.digital.hmpps.crimeportalgateway.application
 
-import com.microsoft.applicationinsights.TelemetryClient
-import com.microsoft.applicationinsights.boot.dependencies.apachecommons.lang3.StringUtils
 import com.microsoft.applicationinsights.telemetry.RequestTelemetry
 import com.microsoft.applicationinsights.web.internal.RequestTelemetryContext
 import com.microsoft.applicationinsights.web.internal.ThreadContext
 import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Condition
-import org.springframework.context.annotation.ConditionContext
-import org.springframework.context.annotation.Conditional
-import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Profile
-import org.springframework.core.type.AnnotatedTypeMetadata
-import org.springframework.lang.NonNull
 import org.springframework.web.context.annotation.RequestScope
 import java.util.Optional
-import java.util.UUID
 
-@Configuration
-class TelemetryConfig {
-
-    @Bean
-    @Conditional(AppInsightKeyAbsentCondition::class)
-    fun getTelemetryClient(): TelemetryClient {
-        return TelemetryClient()
-    }
-
-    private class AppInsightKeyAbsentCondition : Condition {
-        override fun matches(@NonNull context: ConditionContext, @NonNull metadata: AnnotatedTypeMetadata): Boolean {
-            return StringUtils.isEmpty(context.environment.getProperty("application.insights.ikey"))
-        }
-    }
+@Profile(value = ["!test", "!local"])
+class TelemetryConfig : BaseTelemetryConfig() {
 
     @Bean
-    @Profile("!test")
     @RequestScope
     fun requestProperties(): Map<String, String> {
         return Optional.ofNullable(ThreadContext.getRequestTelemetryContext())
@@ -44,7 +22,6 @@ class TelemetryConfig {
 
     @Bean
     fun getOperationId(): () -> String? {
-        return { UUID.randomUUID().toString() }
-        // return { ThreadContext.getRequestTelemetryContext()?.httpRequestTelemetry?.context?.operation?.id }
+        return { ThreadContext.getRequestTelemetryContext()?.httpRequestTelemetry?.context?.operation?.id }
     }
 }
